set the size of all serialized enums

+++ code/bits/Shape.h
@@ -24,7 +24,7 @@
 
 namespace akgr {
 
-  enum class ShapeType : int32_t {
+  enum class ShapeType : uint8_t {
     None,
     Circle,
     Rectangle,
